+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/StockLocationLineServiceSupplychainImpl.java
@@ -21,9 +21,11 @@ import com.axelor.apps.base.db.Product;
 import com.axelor.apps.base.service.app.AppBaseService;
 import com.axelor.apps.stock.db.StockLocation;
 import com.axelor.apps.stock.db.StockLocationLine;
+import com.axelor.apps.stock.db.TrackingNumber;
 import com.axelor.apps.stock.db.repo.StockLocationLineRepository;
 import com.axelor.apps.stock.db.repo.StockMoveLineRepository;
 import com.axelor.apps.stock.db.repo.WapHistoryRepository;
+import com.axelor.apps.stock.service.StockLocationLineService;
 import com.axelor.apps.stock.service.StockLocationLineServiceImpl;
 import com.axelor.apps.stock.service.StockRulesService;
 import com.axelor.apps.supplychain.exception.IExceptionMessage;
@@ -39,19 +41,23 @@ import java.math.BigDecimal;
 @RequestScoped
 public class StockLocationLineServiceSupplychainImpl extends StockLocationLineServiceImpl {
 
+  protected AppSupplychainService appSupplychainService;
+
   @Inject
   public StockLocationLineServiceSupplychainImpl(
       StockLocationLineRepository stockLocationLineRepo,
       StockRulesService stockRulesService,
       StockMoveLineRepository stockMoveLineRepository,
       AppBaseService appBaseService,
-      WapHistoryRepository wapHistoryRepo) {
+      WapHistoryRepository wapHistoryRepo,
+      AppSupplychainService appSupplychainService) {
     super(
         stockLocationLineRepo,
         stockRulesService,
         stockMoveLineRepository,
         appBaseService,
         wapHistoryRepo);
+    this.appSupplychainService = appSupplychainService;
   }
 
   @Override
@@ -84,6 +90,30 @@ public class StockLocationLineServiceSupplychainImpl extends StockLocationLineSe
     BigDecimal availableQty = BigDecimal.ZERO;
     if (stockLocationLine != null) {
       availableQty = stockLocationLine.getCurrentQty().subtract(stockLocationLine.getReservedQty());
+    }
+    return availableQty;
+  }
+
+  @Override
+  public BigDecimal getTrackingNumberAvailableQty(
+      StockLocation stockLocation, TrackingNumber trackingNumber) {
+
+    if (!appSupplychainService.isApp("supplychain")
+        || !appSupplychainService.getAppSupplychain().getManageStockReservation()) {
+      return super.getTrackingNumberAvailableQty(stockLocation, trackingNumber);
+    }
+
+    StockLocationLine detailStockLocationLine =
+        Beans.get(StockLocationLineService.class)
+            .getDetailLocationLine(stockLocation, trackingNumber.getProduct(), trackingNumber);
+
+    BigDecimal availableQty = BigDecimal.ZERO;
+
+    if (detailStockLocationLine != null) {
+      availableQty =
+          detailStockLocationLine
+              .getCurrentQty()
+              .subtract(detailStockLocationLine.getReservedQty());
     }
     return availableQty;
   }